WIP: case-insensitive handling of filenames
This alters file_tree to analyse filenames in a case-insensitive manner.

It needs care - it means some instances of File_set.mem "foo" bar need to
become File_set.find_opt "foo" bar = Some "foo" (i.e. to ensure that the
file has the correct case, e.g. for jbuild and jbuild-ignore).

However, it's also having an effect on the build system - it may be
better to have a function mem_sensitive or something which is guaranteed
to do the lookup in a case-sensitive manner. It also may be desirable to
have mem_ocaml (allowing BuildSystem.ml or buildSystem.ml for module
BuildSystem).

Either way, it's not important enough to have finished off yet...